Modularize project with flask blueprints
