+++ packit_dashboard/config.py
@@ -1,0 +1,5 @@
+from os import getenv
+
+deployment = getenv("DEPLOYMENT", "stg")
+
+API_URL = f"https://{deployment}.packit.dev/api"
